cast to float32 to pass tests in jax021

+++ test/test_distributions.py
@@ -632,7 +632,7 @@ def test_mean_var(jax_dist, sp_dist, params):
     n = 20000 if jax_dist in [dist.LKJ, dist.LKJCholesky] else 200000
     d_jax = jax_dist(*params)
     k = random.PRNGKey(0)
-    samples = d_jax.sample(k, sample_shape=(n,))
+    samples = d_jax.sample(k, sample_shape=(n,)).astype(np.float32)
     # check with suitable scipy implementation if available
     if sp_dist and not _is_batched_multivariate(d_jax):
         d_sp = sp_dist(*params)
